规范化数据集格式并全部测试通过
