+++ ppmat/utils/colored_tqdm.py
@@ -1,17 +1,3 @@
-# Copyright (c) 2026 PaddlePaddle Authors. All Rights Reserved.
-
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-
-#     http://www.apache.org/licenses/LICENSE-2.0
-
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
 from tqdm import tqdm
 import time
 import os;os.system("") #兼容windows
@@ -66,6 +52,10 @@ class ColoredTqdm(tqdm):
         self.end_color = end_color
     
     def get_current_color(self):
+
+        if self.total is None:
+            return "#FFFFFF" 
+
         progress = self.n / self.total if self.total > 0 else 0
         current_rgb = tuple(
             int(start + (end - start) * progress)
@@ -78,7 +68,6 @@ class ColoredTqdm(tqdm):
     
     def update(self, n=1):
         super().update(n)
-        # 使用Rich的真彩色支持
         style = hex_to_ansi(self.get_current_color())
         self.bar_format = f'{{l_bar}}{style}{{bar}}{hex_to_ansi.reset}{{r_bar}}'
         self.refresh()
@@ -86,5 +75,6 @@ class ColoredTqdm(tqdm):
 
 if __name__ == "__main__":
     # 使用示例
-    for i in ColoredTqdm(range(100), desc="🌈 彩虹渐变"):
-        time.sleep(0.1)
+    for i in ColoredTqdm(range(10), desc="🌈 彩虹渐变", leave = False):
+        for j in ColoredTqdm(range(100), desc="🌈 彩虹渐变", leave = False):
+            time.sleep(0.01)
